Added gun and sprint animations with corresponding logic
Added gun and sprint animations with corresponding logic:
1. Ellen can hold the gun only if either Idle or Walking.
2. When Ellen is Sprinting or Jumping she keeps the gun in her right hand.
3. Ellen cannot sprint if she isn't already moving at least a little.
4. Jumping has higher priority over Walking and Sprinting.

Shooting is not implemented yet.
Picking up and dropping a gun is not implemented yet.

If Ellen is able to carry more than one single weapon, she will need a scheme to decide where to store them: for instance, switching from using a gun to using a melee weapon may require to implement the @EllenGunPutAway and the @EllenGunTakeOut animations *and vice versa.)

Finally, currently IsShootPresssed is detected but is not implemented yet, i.e. no bullets/lasers are produced by the gun.