Ellen hold the gun in her right hand
Ellen hold the gun in her right hand:
1. the 'G' key now means 'aim'.
2. The gun has been made as much as possible parallel to the floor.

In the future:
A. IK will make the gun point to a target and Ellen's arms will adapt to that.
B. The gun will actually shoot lasers or bullets.
C. Probably the 'G' key will be eliminated since Ellen will be aiming automatically.