Implemented lateral aiming and perfected a few parameters
Now Ellen's camera moves to the side while aiming.
Also, the crosshair has been parented under the Main Camera: it still needs some logic to detect actual targets and maybe change color when it does so.
Finally, now the camera style is in Combat mode when Ellen is aiming and walking or aiming and shooting, otherwise it is in Adventure mode.