Added ability to switch lateral aim side by either pressing L3 or the F key
Now, when the player presses the L3 button on a controller, or the F key on the keyboard, the lateral aim switches from left to right and vice versa.

Also, the initial lateral aim when the camera is zoomed out has been reduced from 0.2f to zero, so that the camera doesn't more when the lateral aim switching is activated unless the player is also aiming.