Removed mouse cursor while playing
Removed mouse cursor while playing: the mouse is locked to the center of the screen and is invisible.

While in Editor Mode, it is possible to make the cursor re-appear by pressing the ESC key.
While executing a binary of the game, it is possible to make the cursor re-appear by pressing Alt-Tab.