Added Trail Renderer for gun and tweaked shooting animation to be more stable
Added Trail Renderer for gun and tweaked shooting animation to be more stable.

The shooting animation went from frame zero to 35 to frame zero to 3, while the aiming animation has been swapped for the shooting animation: this way there is a lot less wiggling  while Ellen shoots.

The angle of the gun while shooting has been adjusted slightly to be better aligned with the perceived direction of the bullets.

Cleanups:
- The bullet prefab has been deleted.
- The Debug.DrawLine() calls has been commented out.